requirements.txt
Flask==2.1.2
Pillow==9.0.1
ultralytics==8.0.0
tensorflow==2.10.0
